backend/requirements.txt
fastapi
uvicorn
python-docx
google-generativeai
python-multipart
pydantic
python-dotenv
